Move checker test code to test sources
