--- tools/tck-integrity-tests/src/main/scala/org/opencypher/tools/tck/ValidateError.scala
@@ -1,51 +1,0 @@
-/*
- * Copyright (c) 2015-2021 "Neo Technology,"
- * Network Engine for Objects in Lund AB [http://neotechnology.com]
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * Attribution Notice under the terms of the Apache License 2.0
- *
- * This work was created by the collective efforts of the openCypher community.
- * Without limiting the terms of Section 6, any Derivative Work that is not
- * approved by the public consensus process of the openCypher Implementers Group
- * should not be described as “Cypher” (and Cypher® is a registered trademark of
- * Neo4j Inc.) or as "openCypher". Extensions by implementers or prototypes or
- * proposals for change that have been documented or implemented should only be
- * described as "implementation extensions to Cypher" or as "proposed changes to
- * Cypher that are not yet approved by the openCypher community".
- */
-package org.opencypher.tools.tck
-
-import org.opencypher.tools.tck.constants.{TCKErrorDetails, TCKErrorTypes}
-import org.scalatest.funspec.AnyFunSpecLike
-import org.scalatest.matchers.should.Matchers
-
-/**
-  * This function validates a TCK error specification, which consists of three parts: type, phase and detail. Each
-  * of these parts needs to be one of a pre-defined set of constants in order to be valid.
-  */
-trait ValidateError extends AnyFunSpecLike with Matchers {
-
-  def validateError(typ: String, phase: String, detail: String): Unit = {
-    it("has valid type") {
-      TCKErrorTypes.ALL should contain (typ)
-    }
-    it("has valid phase") {
-      Set("runtime", "compile time") should contain (phase)
-    }
-    it("has valid detail") {
-      TCKErrorDetails.ALL should contain (detail)
-    }
-  }
-}
